+++ src/main/java/dev/shadowsoffire/apotheosis/socket/gem/bonus/GemBonus.java
@@ -20,6 +20,7 @@ import dev.shadowsoffire.apotheosis.socket.gem.bonus.special.FrozenDropsBonus;
 import dev.shadowsoffire.apotheosis.socket.gem.bonus.special.LeechBlockBonus;
 import dev.shadowsoffire.apotheosis.socket.gem.bonus.special.MageSlayerBonus;
 import dev.shadowsoffire.apotheosis.socket.gem.bonus.special.OmneticBonus;
+import dev.shadowsoffire.apotheosis.socket.gem.bonus.special.RadialBonus;
 import dev.shadowsoffire.apothic_attributes.modifiers.StackAttributeModifiersEvent;
 import dev.shadowsoffire.placebo.codec.CodecMap;
 import dev.shadowsoffire.placebo.codec.CodecProvider;
@@ -273,6 +274,7 @@ public abstract class GemBonus implements CodecProvider<GemBonus> {
         register("mob_effect", MobEffectBonus.CODEC);
         register("frozen_drops", FrozenDropsBonus.CODEC);
         register("omnetic", OmneticBonus.CODEC);
+        register("radial", RadialBonus.CODEC);
     }
 
     protected static <T extends GemBonus> App<RecordCodecBuilder.Mu<T>, GemClass> gemClass() {