+++ src/main/java/dev/shadowsoffire/apotheosis/socket/gem/bonus/special/RadialBonus.java
@@ -1,0 +1,91 @@
+package dev.shadowsoffire.apotheosis.socket.gem.bonus.special;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+import com.mojang.serialization.Codec;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+
+import dev.shadowsoffire.apotheosis.Apoth.Components;
+import dev.shadowsoffire.apotheosis.socket.SocketHelper;
+import dev.shadowsoffire.apotheosis.socket.gem.GemClass;
+import dev.shadowsoffire.apotheosis.socket.gem.GemInstance;
+import dev.shadowsoffire.apotheosis.socket.gem.GemView;
+import dev.shadowsoffire.apotheosis.socket.gem.Purity;
+import dev.shadowsoffire.apotheosis.socket.gem.bonus.GemBonus;
+import dev.shadowsoffire.apotheosis.util.RadialUtil;
+import dev.shadowsoffire.apotheosis.util.RadialUtil.RadialData;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.neoforged.neoforge.common.util.AttributeTooltipContext;
+import net.neoforged.neoforge.event.level.BlockEvent;
+
+public class RadialBonus extends GemBonus {
+
+    public static final Codec<RadialBonus> CODEC = RecordCodecBuilder.create(inst -> inst
+        .group(
+            gemClass(),
+            Purity.mapCodec(RadialData.CODEC).fieldOf("values").forGetter(a -> a.values))
+        .apply(inst, RadialBonus::new));
+
+    protected final Map<Purity, RadialData> values;
+
+    public RadialBonus(GemClass gemClass, Map<Purity, RadialData> values) {
+        super(gemClass);
+        this.values = values;
+    }
+
+    @Override
+    public Codec<? extends GemBonus> getCodec() {
+        return CODEC;
+    }
+
+    @Override
+    public boolean supports(Purity purity) {
+        return this.values.containsKey(purity);
+    }
+
+    @Override
+    public Component getSocketBonusTooltip(GemView gem, AttributeTooltipContext ctx) {
+        RadialData data = this.values.get(gem.purity());
+        return Component.translatable("affix.apotheosis:breaker/effect/radial.desc", data.x(), data.y()).withStyle(ChatFormatting.YELLOW);
+    }
+
+    // EventPriority.LOW
+    public static void onBreak(BlockEvent.BreakEvent e) {
+        Player player = e.getPlayer();
+        ItemStack tool = player.getMainHandItem();
+        Level world = player.level();
+        if (!world.isClientSide && tool.has(Components.SOCKETED_GEMS)) {
+            GemInstance inst = SocketHelper.getGems(tool).streamValidGems().filter(g -> g.getBonus().orElse(null) instanceof RadialBonus).findFirst().orElse(null);
+            if (inst != null && inst.isValid()) {
+                RadialData data = ((RadialBonus) inst.getBonus().get()).values.get(inst.purity());
+                RadialUtil.attemptRadialMining(e, data);
+            }
+        }
+    }
+
+    public static Builder builder() {
+        return new Builder();
+    }
+
+    public static class Builder extends GemBonus.Builder {
+
+        private final Map<Purity, RadialData> values = new LinkedHashMap<>();
+
+        public Builder value(Purity rarity, int x, int y, int xOff, int yOff) {
+            RadialData data = new RadialData(x, y, xOff, yOff);
+            this.values.put(rarity, data);
+            return this;
+        }
+
+        public RadialBonus build(GemClass gClass) {
+            return new RadialBonus(gClass, this.values);
+        }
+
+    }
+
+}